release: cleanup and make everything namespace driven
This introduces namespaced custom properties allowing users of the library to override the defaults of Minimalist.
This aligns with the documentation at https://minimalist-docs.netlify.app/
Sets up the library to drive the typography through `calc` and possible some use of `clamp`
Updates the README to officially introduce the project.
Many other small but meaningful updates.